requirements.txt
aiohttp==3.13.2
Pillow==12.0.0
python-dotenv==1.2.1
python-socketio==5.15.1
websockets==15.0.1
